Move SystemMessageMailable rendering
